Let the nearby actors chase the player
Using AStar of ROT, let all actors query for a path to the player, and
only the near ones will chase down.

There is an issue when collision with player happens.

There may be a bug in free tiles. Maybe we don't need to keep them in
state at all.

And there are a bunch of duplicate codes. Smell!