Change the menu scene to be like a chat room
Probably showing an offer by an anonymous agent, taking the player to a
training mission to see if the player qualifies to join organization.

This Chat component would probably be reused through out the game as a
way to communicate missions between the organization and the player, as
well as other random jobs.

Most of the Chat component are random content to make it "animating".