Change the menu screen to be less menu and more interactions
Instead of using traditional menu and buttons to start game, display an
IRC client showing the player is connecting to a special channel. Some
time later an anonymous agent contacts the player for a job, w/ specific
instructions to guide the player to the actual combat.

Chat happen in different tabs. Player is generated a random nickname on
joining, but can change via `/nick newnickname`. `/msg NicServ identify`
does not function yet.

A random list of users is generated, and random users join and leave to
show a little bit of dynamic.