End game by reaching the goal
When the player reaches the goal, the game returns the player back
to the chat room, with the current agent offering a new job.

Ideally, the player should still be able to play the current game,
until he closes the hacking tool, so that he can collects extra
files and (maybe) experience.