Process Manager stats and damage calculation
Depends on the selected OS, the player will have initial memory usage
and damage differently. When on the map, player has a chance to attack
first and only gets damage when the enemy is still alive.

Player's memory usage is update on the Process Manager. Every hit by
enemy will increase the memory usage of the editor. If it passes limit,
the game ends.